Todo API/pyproject.toml
[project]
name = "todo-api"
version = "0.1.0"
description = "TODO API built with FastAPI + PostgreSQL"
authors = [
    {name = "Jamezfn", email = "jamesmakachia19@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.4,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "app"}]